feat: export PDF, MoM comparison report, dashboard date-range filter
- Add Export PDF button on expenses listing (barryvdh/laravel-dompdf,
  carries active filters, A4 portrait with category colour pills and totals row)
- Add month-over-month comparison report (/reports/mom-comparison):
  per-category this vs last month, % change with up/down arrow badges,
  "New" pill for first-time categories, totals footer row
- Add dashboard date-range filter: From/To date inputs + preset chips
  (This Month / Last Month / Last 30d / Last 90d / This Year); stats,
  donut chart, and top-categories bar all reflect the selected period
- Fix category badge style2 attribute bug in expenses listing
- Fix @selected type mismatch on expense create/edit forms
- Fix TypeError from string→int cast on expense edit/delete/show
- Consistent table row text styling across all listing views
- Add sidebar nav entry for vs. Last Month report

+++ resources/views/expenses/create.blade.php
@@ -48,7 +48,7 @@
                             class="select w-full @error('category_id') border-red-400 @enderror">
                         <option value="">Select a category</option>
                         @foreach ($categories as $category)
-                            <option value="{{ $category->id }}" @selected(old('category_id') === $category->id)>
+                            <option value="{{ $category->id }}" @selected((int) old('category_id') === $category->id)>
                                 {{ $category->name }}
                             </option>
                         @endforeach
